Move ast module files to root project
