--- ast/build.gradle
@@ -1,15 +1,0 @@
-plugins {
-    id 'groovy'
-}
-
-group 'com.github.breadmoirai'
-version '2.3.0'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation gradleApi()
-    testImplementation 'org.codehaus.groovy:groovy-test:3.0.9'
-}
